+++ wp-content/plugins/advanced-responsive-video-embedder/shared/functions-shared.php
@@ -36,7 +36,23 @@ function arve_get_options_defaults( $section ) {
  */
 function arve_get_options() {
 
-	$options               = wp_parse_args( get_option( 'arve_options_main',       array() ), arve_get_options_defaults( 'main' ) );
+	$options = wp_parse_args( get_option( 'arve_options_main', array() ), arve_get_options_defaults( 'main' ) );
+
+	$supported_modes = arve_get_supported_modes();
+
+	# legacy mode name
+	if ( 'thumbnail' == $options['mode'] ) {
+
+		$options['mode'] = 'lazyload';
+		update_option( 'arve_options_main', $options );
+	}
+
+	if( ! in_array( $options['mode'], array( 'normal', 'lazyload', 'lazyload-lightbox', 'link-lightbox' ) ) ) {
+
+		$options['mode'] = 'lazyload';
+		update_option( 'arve_options_main', $options );
+	}
+
 	$options['shortcodes'] = wp_parse_args( get_option( 'arve_options_shortcodes', array() ), arve_get_options_defaults( 'shortcodes' ) );
 	$options['params']     = wp_parse_args( get_option( 'arve_options_params',     array() ), arve_get_options_defaults( 'params' ) );
 
@@ -727,13 +743,14 @@ function arve_get_host_properties() {
 			),
 		),
 		'vimeo' => array(
-			'regex'          => $s . 'vimeo\.com/(?:(?:channels/[a-z]+/)|(?:groups/[a-z]+/videos/))?([0-9]+)',
+			'regex'          => 'https?://(?:player\.)?vimeo\.com/(?:(?:video/)|(?:channels/[a-z]+/)|(?:groups/[a-z]+/videos/))?([0-9]+)',
 			'embed_url'      => 'https://player.vimeo.com/video/%s',
 			'default_params' => 'html5=1&title=1&byline=0&portrait=0',
 			'auto_thumbnail' => true,
 			'auto_title'     => true,
 			'tests' => array(
-				array( 'url' => 'https://vimeo.com/124400795', 'id' => 124400795 ),
+				array( 'url' => 'https://vimeo.com/124400795',        'id' => 124400795 ),
+				array( 'url' => 'https://player.vimeo.com/124400795', 'id' => 124400795 ),
 			),
 			'query_argss' => array(
 				'autoplay'  => array( 'bool', __( 'Autoplay', ARVE_SLUG ) ),