+++ wp-content/plugins/contact-form-7-mailchimp-extension/cf7-mch-ext.php
@@ -7,7 +7,7 @@ Author: Renzo Johnson
 Author URI: http://renzojohnson.com
 Text Domain: contact-form-7
 Domain Path: /languages/
-Version: 0.4.36
+Version: 0.4.37
 */
 
 /*  Copyright 2013-2017 Renzo Johnson (email: renzojohnson at gmail.com)
@@ -27,7 +27,7 @@ Version: 0.4.36
     Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 */
 
-define( 'SPARTAN_MCE_VERSION', '0.4.36' );
+define( 'SPARTAN_MCE_VERSION', '0.4.37' );
 define( 'SPARTAN_MCE_PLUGIN_BASENAME', plugin_basename( __FILE__ ) );
 define( 'SPARTAN_MCE_PLUGIN_NAME', trim( dirname( SPARTAN_MCE_PLUGIN_BASENAME ), '/' ) );
 define( 'SPARTAN_MCE_PLUGIN_DIR', untrailingslashit( dirname( __FILE__ ) ) );
@@ -39,28 +39,12 @@ require_once( SPARTAN_MCE_PLUGIN_DIR . '/lib/mailchimp.php' );
 function mc_meta_links( $links, $file ) {
     if ( $file === 'contact-form-7-mailchimp-extension/cf7-mch-ext.php' ) {
         $links[] = '<a href="'.MCE_URL.'" target="_blank" title="Documentation">Documentation</a>';
-        $links[] = '<a href="'.MCE_URL.'" target="_blank" title="Starter Guide">Starter Guide</a>';
-        $links[] = '<a href="//www.paypal.com/cgi-bin/webscr?cmd=_s-xclick&hosted_button_id=UD6FL5SCU5K8W" target="_blank" title="Donate"><strong>Donate</strong></a>';
+        // $links[] = '<a href="'.MCE_URL.'" target="_blank" title="Starter Guide">Starter Guide</a>';
+        $links[] = '<a href="//www.paypal.me/renzojohnson" target="_blank" title="Donate"><strong>Donate</strong></a>';
     }
     return $links;
 }
 add_filter( 'plugin_row_meta', 'mc_meta_links', 10, 2 );
-
-
-function mc_get_latest_item(){
-    $args = array(
-            'post_type'         => 'wpcf7_contact_form',
-            'posts_per_page'    => -1,
-            'fields'            => 'ids',
-        );
-    // Get Highest Value from CF7Forms
-    $form = max(get_posts($args));
-    $out = '';
-    if (!empty($form)) {
-        $out .= $form;
-    }
-    return $out;
-}
 
 
 function mc_settings_link( $links ) {
@@ -72,18 +56,9 @@ function mc_settings_link( $links ) {
 
 
 function mc_after_setup_theme() {
-     add_filter('plugin_action_links_' . plugin_basename(__FILE__), 'mc_settings_link');
+     add_filter('plugin_action_links_' . SPARTAN_MCE_PLUGIN_BASENAME, 'mc_settings_link');
 }
 add_action ('after_setup_theme', 'mc_after_setup_theme');
-
-
-
-function mce_act_redirect( $plugin ) {
-  if( $plugin == SPARTAN_MCE_PLUGIN_BASENAME ) {
-    exit( wp_redirect( admin_url( 'admin.php?page=wpcf7&post='.mc_get_latest_item().'&active-tab=4' ) ) );
-  }
-}
-add_action( 'activated_plugin', 'mce_act_redirect' );